Add go-checks reusable workflow (#1)
The gofumpt and go.mod tidy checks were tested on a repository.

I also added some usage instructions for the JIRA sync, and did
a simple test, though I was not able to test the whole action
completely yet.

For backports, I wrote something that might work, but when
I tried to test it, it always failed with:

```
2023/03/01 23:18:39 backport-assistant error: error while applying
environment option in New: error while authenticating with GitHub:
GET https://api.github.com/user: 403 Resource not accessible by integration []
```

I don't think we use the backport flow that often for plugins, so it
may not get used. (And even so, it is only a single step, so it isn't
terrible to copy-paste, if we can figure out the permissions issue).

I did not do add a reusable check for the changelog, as I'm
not sure exactly what it would do. Should it check that the changelog
is updated in ever PR?

I also did not add a reusable check for security-scanner, as that
is available as a action step with minimal configuration.